skills/ai-agents/context-engineering/requirements.txt
pyairtable==2.3.3
python-dotenv==1.0.1
requests==2.32.3
notion-client==2.2.1
